rename shot-explorer/main to shot-explorer/setup
it runs in a renderer and not in a main process
