+++ src/js/windows/shot-explorer/setup.js
